Interpreting unsigned char as int8_t

+++ src/main/java/com/frostwire/jlibtorrent/AnnounceEntry.java
@@ -75,7 +75,7 @@ public final class AnnounceEntry {
      *
      * @return
      */
-    public short tier() {
+    public int tier() {
         return e.getTier();
     }
 }
